Move ruleRoutine to ruleInfo file
